big file support and tests!

+++ src/hobbits-core/frame.h
@@ -10,11 +10,11 @@
 class HOBBITSCORESHARED_EXPORT Frame : public Range
 {
 public:
-    Frame(QSharedPointer<const BitArray> bits, int start, int end);
+    Frame(QSharedPointer<const BitArray> bits, qint64 start, qint64 end);
     Frame(QSharedPointer<const BitArray> bits, Range range);
     Frame(const Frame &other) = default;
 
-    bool at(int i) const;
+    bool at(qint64 i) const;
 
 private:
     QSharedPointer<const BitArray> m_bits;
